move `parse_fingerprint_entry_as_variable` to fingerprinting utils module
