+++ src/sentry/grouping/utils.py
@@ -9,24 +9,16 @@ from uuid import UUID
 from django.utils.encoding import force_bytes
 
 from sentry.db.models.fields.node import NodeData
-from sentry.grouping.fingerprinting.utils import DEFAULT_FINGERPRINT_VARIABLE, _fingerprint_var_re
+from sentry.grouping.fingerprinting.utils import (
+    DEFAULT_FINGERPRINT_VARIABLE,
+    _fingerprint_var_re,
+    parse_fingerprint_entry_as_variable,
+)
 from sentry.stacktraces.processing import get_crash_frame_from_event_data
 from sentry.utils.safe import get_path
 
 if TYPE_CHECKING:
     from sentry.grouping.component import ExceptionGroupingComponent
-
-
-def parse_fingerprint_entry_as_variable(entry: str) -> str | None:
-    """
-    Determine if the given fingerprint entry is a variable, and if it is, return its key (that is,
-    extract the variable name from a variable string of the form "{{ var_name }}"). If the given
-    entry isn't the correct form to be a variable, return None.
-    """
-    match = _fingerprint_var_re.match(entry)
-    if match is not None and match.end() == len(entry):
-        return match.group(1)
-    return None
 
 
 def is_default_fingerprint_var(value: str) -> bool:
